Add blanket Resolver<K> impls for &T and &mut T
